backend/requirements.txt
openbb
fastapi
uvicorn
pydantic
requests
pyyaml
plotly
pandas
python-dotenv
